chore: move generateAccounts to seerate file (#1623)
