+++ apps/console-lite-e2e/src/support/mocks/generate-accounts.ts
@@ -1,7 +1,6 @@
 import merge from 'lodash/merge';
 import type { AccountsQuery } from '@vegaprotocol/accounts';
-import type { AssetsQuery } from '@vegaprotocol/assets';
-import { AccountType, Schema as Types } from '@vegaprotocol/types';
+import { AccountType } from '@vegaprotocol/types';
 import type { PartialDeep } from 'type-fest';
 
 export const generateAccounts = (
@@ -75,47 +74,4 @@ export const generateAccounts = (
     },
   };
   return merge(defaultAccounts, override);
-};
-
-export const generateAssets = (override?: PartialDeep<AssetsQuery>) => {
-  const defaultAssets: AssetsQuery = {
-    assetsConnection: {
-      edges: [
-        {
-          node: {
-            id: 'asset-id',
-            symbol: 'tEURO',
-            decimals: 5,
-            name: 'Euro',
-            source: {},
-            quantum: '',
-            status: Types.AssetStatus.STATUS_ENABLED,
-          },
-        },
-        {
-          node: {
-            id: 'asset-id-2',
-            symbol: 'tDAI',
-            decimals: 5,
-            name: 'DAI',
-            source: {},
-            quantum: '',
-            status: Types.AssetStatus.STATUS_ENABLED,
-          },
-        },
-        {
-          node: {
-            id: 'asset-0',
-            symbol: 'AST0',
-            decimals: 5,
-            name: 'Asto',
-            source: {},
-            quantum: '',
-            status: Types.AssetStatus.STATUS_ENABLED,
-          },
-        },
-      ],
-    },
-  };
-  return merge(defaultAssets, override);
 };
